Closes #28 : Segregation of ssh , validation and main method
